requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-dotenv==1.0.0
rapidfuzz==3.6.1
psycopg-binary==3.2.9
psycopg~=3.2.9